refactor: decompose ClaudeBackend
